requirements.txt
requests
python-dotenv
pytest
schedule
win10toast
fastapi
uvicorn
discord.py
slack_bolt

# Optional extras
streamlit
plyer
pyttsx3
SpeechRecognition
pyaudio
PyQt5
textual
